+++ apps/swaps/src/types.ts
@@ -1,0 +1,5 @@
+import type { SwapKitCore } from '@coinmasters/core';
+
+export type WalletDataType = Awaited<
+  ReturnType<InstanceType<typeof SwapKitCore>['getWalletByChain']>
+>;