+++ packages/swapkit/core/src/client/index.ts
@@ -497,9 +497,9 @@ export class SwapKitCore<T = ''> {
 
   extend = ({ wallets, config, apis = {}, rpcUrls = {} }: ExtendParams<T>) => {
     try {
-      console.log("wallets: ",wallets)
+      console.log('wallets: ', wallets);
       wallets.forEach((wallet) => {
-        console.log("wallet: ",wallet)
+        console.log('wallet: ', wallet);
         // @ts-expect-error ANCHOR - Not Worth
         this[wallet.connectMethodName] = wallet.connect({
           addChain: this.#addConnectedChain,