+++ packages/toolboxes/cosmos/src/types.ts
@@ -1,7 +1,7 @@
-import type { StdFee } from '@cosmjs/amino';
-import type { OfflineDirectSigner } from '@cosmjs/proto-signing';
 import type { AssetValue } from '@coinmasters/helpers';
 import type { ChainId, FeeOption } from '@coinmasters/types';
+import type { StdFee } from '@cosmjs/amino';
+import type { OfflineDirectSigner } from '@cosmjs/proto-signing';
 
 import type {
   BinanceToolboxType,