+++ packages/wallets/wc/src/walletconnect.ts
@@ -1,7 +1,11 @@
+import type {
+  BaseCosmosToolboxType,
+  DepositParam,
+  TransferParams,
+} from '@coinmasters/toolbox-cosmos';
+import { ApiUrl, Chain, ChainId, WalletOption } from '@coinmasters/types';
 import type { StdSignDoc } from '@cosmjs/amino';
 import type { TxBodyEncodeObject } from '@cosmjs/proto-signing';
-import type { BaseCosmosToolboxType, DepositParam, TransferParams } from '@coinmasters/toolbox-cosmos';
-import { ApiUrl, Chain, ChainId, WalletOption } from '@coinmasters/types';
 import type { WalletConnectModalSign } from '@walletconnect/modal-sign-html';
 import type { SessionTypes, SignClientTypes } from '@walletconnect/types';
 import { SignMode } from 'cosmjs-types/cosmos/tx/signing/v1beta1/signing.js';