+++ packages/swapkit/sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @coinmasters/sdk
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`96ccc78`](https://github.com/thorswap/SwapKit/commit/96ccc7869bd4c6bb99e0ba0a3863d08a08c2fdcd), [`3a230b5`](https://github.com/thorswap/SwapKit/commit/3a230b5265a164d64f29838ceebba5b944fce532)]:
+  - @coinmasters/wallet-ledger@1.0.0
+  - @coinmasters/wallet-trezor@1.0.0
+  - @coinmasters/toolbox-evm@1.0.0
+  - @coinmasters/toolbox-cosmos@1.0.0
+  - @coinmasters/core@1.0.0
+  - @coinmasters/wallet-evm-extensions@1.0.0
+  - @coinmasters/wallet-keystore@1.0.0
+  - @coinmasters/wallet-okx@1.0.0
+  - @coinmasters/wallet-wc@1.0.0
+  - @coinmasters/wallet-xdefi@1.0.0
+  - @coinmasters/wallet-keplr@1.0.0
+  - @coinmasters/api@1.0.0
+  - @coinmasters/types@1.0.0
+  - @coinmasters/toolbox-utxo@1.0.0
+
 ## 1.0.0-rc.2
 
 ### Patch Changes