+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coinmasters/toolbox-cosmos
 
+## 1.0.0
+
+### Patch Changes
+
+- [`3a230b5`](https://github.com/thorswap/SwapKit/commit/3a230b5265a164d64f29838ceebba5b944fce532) Thanks [@chillios-ts](https://github.com/chillios-ts)! - Fix Maya chain connection
+
+- Updated dependencies []:
+  - @coinmasters/api@1.0.0
+  - @coinmasters/helpers@1.0.0
+  - @coinmasters/types@1.0.0
+
 ## 1.0.0-rc.1
 
 ### Patch Changes