+++ packages/toolboxes/evm/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/toolbox-evm
 
+## 1.0.0
+
+### Patch Changes
+
+- [#352](https://github.com/thorswap/SwapKit/pull/352) [`96ccc78`](https://github.com/thorswap/SwapKit/commit/96ccc7869bd4c6bb99e0ba0a3863d08a08c2fdcd) Thanks [@towanTG](https://github.com/towanTG)! - Changes to EVM Signer to fix calls for ethers v6
+
+- Updated dependencies []:
+  - @coinmasters/helpers@1.0.0
+  - @coinmasters/types@1.0.0
+
 ## 1.0.0-rc.1
 
 ### Patch Changes