+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @coinmasters/wallet-evm-extensions
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`96ccc78`](https://github.com/thorswap/SwapKit/commit/96ccc7869bd4c6bb99e0ba0a3863d08a08c2fdcd)]:
+  - @coinmasters/toolbox-evm@1.0.0
+  - @coinmasters/types@1.0.0
+
 ## 1.0.0-rc.1
 
 ### Patch Changes