+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/wallet-keplr
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`3a230b5`](https://github.com/thorswap/SwapKit/commit/3a230b5265a164d64f29838ceebba5b944fce532)]:
+  - @coinmasters/toolbox-cosmos@1.0.0
+  - @coinmasters/helpers@1.0.0
+  - @coinmasters/types@1.0.0
+
 ## 1.0.0-rc.1
 
 ### Patch Changes