+++ packages/wallets/ledger/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @coinmasters/wallet-ledger
 
+## 1.0.0
+
+### Patch Changes
+
+- [#352](https://github.com/thorswap/SwapKit/pull/352) [`96ccc78`](https://github.com/thorswap/SwapKit/commit/96ccc7869bd4c6bb99e0ba0a3863d08a08c2fdcd) Thanks [@towanTG](https://github.com/towanTG)! - Changes to EVM Signer to fix calls for ethers v6
+
+- Updated dependencies [[`96ccc78`](https://github.com/thorswap/SwapKit/commit/96ccc7869bd4c6bb99e0ba0a3863d08a08c2fdcd), [`3a230b5`](https://github.com/thorswap/SwapKit/commit/3a230b5265a164d64f29838ceebba5b944fce532)]:
+  - @coinmasters/toolbox-evm@1.0.0
+  - @coinmasters/toolbox-cosmos@1.0.0
+  - @coinmasters/helpers@1.0.0
+  - @coinmasters/types@1.0.0
+  - @coinmasters/toolbox-utxo@1.0.0
+
 ## 1.0.0-rc.2
 
 ### Patch Changes